requirements.txt


# PostgreSQL
psycopg[binary]

# RabbitMQ
pika==1.3.2

# Outlook integration (Windows only)
pywin32==311

# HTML parsing
beautifulsoup4==4.12.2

# Testing
pytest>=8.2